Add option to ignore specific named 'unused' imports.
If ignored imports are found, they are reported with a warning but grunt will continue.

Provides finer-grain control than using the force: true option, which basically ignores all unused imports. It allows you to ignore certain unused imports which you know are valid (for example the 'react' import in JSX files, where the import is not referenced in the JSX code but will be referenced in the transpiled js code, and is therefore required) but break the build on others.